Do dia 11/7

+++ Fatorial/src/br/com/puts/FatorialUm.java
@@ -2,7 +2,7 @@ package br.com.puts;
 
 public class FatorialUm {
 	public static void main(String[] args) {
-		long fatorial = 1;
+		double fatorial = 1;
 		int x = 20;
 		for (int n = 1; n <= x; n++) {
 			fatorial = n * fatorial;
